# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Books Service","description":"HTTP service for managing books","version":"1.0"},"servers":[{"url":"http://localhost:8080"}],"paths":{"/books":{"get":{"tags":["books"],"summary":"findAll books","description":"FindAll Books","operationId":"books#findAll","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookCollection"},"example":[{"author":"Quia molestiae quibusdam eligendi perspiciatis.","book_cover_bytes":"TW9sZXN0aWFlIGFsaXF1aWQgdmVsaXQgdm9sdXB0YXMu","book_cover_file_name":"Modi quis omnis et odio est quo.","book_cover_file_type":"Sed animi.","id":8346939551506138778,"published_at":"Ut sint accusamus.","title":"Temporibus rerum corrupti magnam praesentium a."},{"author":"Quia molestiae quibusdam eligendi perspiciatis.","book_cover_bytes":"TW9sZXN0aWFlIGFsaXF1aWQgdmVsaXQgdm9sdXB0YXMu","book_cover_file_name":"Modi quis omnis et odio est quo.","book_cover_file_type":"Sed animi.","id":8346939551506138778,"published_at":"Ut sint accusamus.","title":"Temporibus rerum corrupti magnam praesentium a."}]}}}}},"post":{"tags":["books"],"summary":"create books","description":"Create a new book.","operationId":"books#create","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateRequestBody"},"example":{"author":"Repellendus itaque et.","book_cover":{"bytes":"SW4gcmVtIG5vbi4=","name":"Architecto sapiente nesciunt aut repellendus rerum quia.","type":"Rerum enim quibusdam qui qui hic quia."},"published_at":"2024-01-27","title":"Cum quia."}}}},"responses":{"201":{"description":"Created response.","content":{"application/json":{"schema":{"type":"string","example":"Tempora labore quam maiores ut nihil."},"example":"Perspiciatis assumenda quibusdam enim."}}}}}},"/books/{id}":{"delete":{"tags":["books"],"summary":"delete books","description":"Delete a book.","operationId":"books#delete","parameters":[{"name":"id","in":"path","description":"Book ID","required":true,"schema":{"type":"integer","description":"Book ID","example":7333511080261442393,"format":"int64"},"example":5723373541677809871}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"string","example":"Ab pariatur."},"example":"Consectetur temporibus ullam expedita."}}},"404":{"description":"NotFound: Not Found response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"get":{"tags":["books"],"summary":"findById books","description":"Find book by id","operationId":"books#findById","parameters":[{"name":"id","in":"path","description":"Book ID","required":true,"schema":{"type":"integer","description":"Book ID","example":8711181424041302409,"format":"int64"},"example":6319500947888338300}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"},"example":{"author":"Quia molestiae quibusdam eligendi perspiciatis.","book_cover_bytes":"TW9sZXN0aWFlIGFsaXF1aWQgdmVsaXQgdm9sdXB0YXMu","book_cover_file_name":"Modi quis omnis et odio est quo.","book_cover_file_type":"Sed animi.","id":8346939551506138778,"published_at":"Ut sint accusamus.","title":"Temporibus rerum corrupti magnam praesentium a."}}}},"404":{"description":"NotFound: Not Found response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"put":{"tags":["books"],"summary":"update books","description":"Update a book.","operationId":"books#update","parameters":[{"name":"id","in":"path","description":"Book ID","required":true,"schema":{"type":"integer","description":"Book ID","example":3426596595496711312,"format":"int64"},"example":6845661201716045276}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UpdateRequestBody"},"example":{"author":"Tenetur nemo amet iste saepe molestias est.","book_cover":{"bytes":"SW4gcmVtIG5vbi4=","name":"Architecto sapiente nesciunt aut repellendus rerum quia.","type":"Rerum enim quibusdam qui qui hic quia."},"published_at":"2023-12-31","title":"Consequatur sapiente aut occaecati."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"string","example":"Dolorem enim porro qui."},"example":"Natus quidem."}}},"404":{"description":"NotFound: Not Found response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}}},"components":{"schemas":{"Book":{"type":"object","properties":{"author":{"type":"string","description":"Author of book","example":"Culpa in qui."},"book_cover_bytes":{"type":"string","description":"Image File base64","example":"T2ZmaWNpYSBleGVyY2l0YXRpb25lbS4=","format":"binary"},"book_cover_file_name":{"type":"string","description":"Book cover file name","example":"Omnis eos."},"book_cover_file_type":{"type":"string","description":"Book cover file type","example":"Nesciunt aperiam veritatis eligendi porro natus quos."},"id":{"type":"integer","description":"ID of book","example":6149561257650898521,"format":"int64"},"published_at":{"type":"string","description":"Published date of the book","example":"Sit ipsam."},"title":{"type":"string","description":"Title of book","example":"Atque non consequatur totam culpa."}},"description":"A book payload","example":{"author":"Aperiam laborum quia accusamus dolorum omnis.","book_cover_bytes":"RG9sb3JlcyByZWN1c2FuZGFlIHJlaWNpZW5kaXMu","book_cover_file_name":"Consequatur ipsam at quia.","book_cover_file_type":"A aut asperiores.","id":5237505862077042108,"published_at":"Unde et ducimus magnam nemo porro aut.","title":"Et autem nemo commodi."},"required":["id","title","author","published_at"]},"BookCollection":{"type":"array","items":{"$ref":"#/components/schemas/Book"},"example":[{"author":"Quia molestiae quibusdam eligendi perspiciatis.","book_cover_bytes":"TW9sZXN0aWFlIGFsaXF1aWQgdmVsaXQgdm9sdXB0YXMu","book_cover_file_name":"Modi quis omnis et odio est quo.","book_cover_file_type":"Sed animi.","id":8346939551506138778,"published_at":"Ut sint accusamus.","title":"Temporibus rerum corrupti magnam praesentium a."},{"author":"Quia molestiae quibusdam eligendi perspiciatis.","book_cover_bytes":"TW9sZXN0aWFlIGFsaXF1aWQgdmVsaXQgdm9sdXB0YXMu","book_cover_file_name":"Modi quis omnis et odio est quo.","book_cover_file_type":"Sed animi.","id":8346939551506138778,"published_at":"Ut sint accusamus.","title":"Temporibus rerum corrupti magnam praesentium a."},{"author":"Quia molestiae quibusdam eligendi perspiciatis.","book_cover_bytes":"TW9sZXN0aWFlIGFsaXF1aWQgdmVsaXQgdm9sdXB0YXMu","book_cover_file_name":"Modi quis omnis et odio est quo.","book_cover_file_type":"Sed animi.","id":8346939551506138778,"published_at":"Ut sint accusamus.","title":"Temporibus rerum corrupti magnam praesentium a."}]},"CreateRequestBody":{"type":"object","properties":{"author":{"type":"string","description":"Author of the book","example":"Animi doloremque occaecati natus quo facilis eum."},"book_cover":{"$ref":"#/components/schemas/File"},"published_at":{"type":"string","description":"Published date of the book","example":"2024-01-27"},"title":{"type":"string","description":"Title of the book","example":"Ad ipsum quia doloribus."}},"example":{"author":"Iusto possimus impedit.","book_cover":{"bytes":"SW4gcmVtIG5vbi4=","name":"Architecto sapiente nesciunt aut repellendus rerum quia.","type":"Rerum enim quibusdam qui qui hic quia."},"published_at":"2024-01-27","title":"Quis amet ea cum id."},"required":["title","author","published_at"]},"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"File":{"type":"object","properties":{"bytes":{"type":"string","example":"UXVpcyB2b2x1cHRhcyBjb25zZXF1YXR1ciBtYWlvcmVzIHV0IGFkIGRvbG9yLg==","format":"binary"},"name":{"type":"string","example":"Consectetur libero sunt."},"type":{"type":"string","example":"Iusto autem praesentium aperiam."}},"description":"Image Upload Type","example":{"bytes":"RGljdGEgdmVsLg==","name":"Officiis ipsam.","type":"Voluptas qui."}},"UpdateRequestBody":{"type":"object","properties":{"author":{"type":"string","description":"Author of the book","example":"Illum fugit."},"book_cover":{"$ref":"#/components/schemas/File"},"published_at":{"type":"string","description":"Published date of the book","example":"2023-12-31"},"title":{"type":"string","description":"Title of the book","example":"Consequatur est voluptas fugit id."}},"example":{"author":"Aliquam omnis.","book_cover":{"bytes":"SW4gcmVtIG5vbi4=","name":"Architecto sapiente nesciunt aut repellendus rerum quia.","type":"Rerum enim quibusdam qui qui hic quia."},"published_at":"2023-12-31","title":"Error quisquam explicabo sed minima."}}}},"tags":[{"name":"books","description":"Books Service"}]}